Update bundle dependencies (+15 squashed commits) Squashed commits: [7bd24ce] Change version flag
Previously, the version flag for the builder was -t, which was
confusing.  This change makes it -v and --version.
[2120273] Add new JRE

[#52036487]
[3bb0a56] Added note about AWS credentials
Previously, there was no noticed about how to obtain AWS credentials.
This change adds that information.

[#51905291]
[79794e8] Index contains full URIs

Previously, the index only contained the relative URI to an artifact.
This change ensures that the index now contains the full URI to the
artifact.

[#51641191]
[270c48f] Change key structure

As part of the renaming of the JRE bucket, the key structure that the
JREs are upload to, must change.  This change does that work.

[#51641279]
[f7489b0] Added README.md

Previously the repository did not have a README.md file.  This change
introduces that file and adds content for how to run the script, along
with a listing of the values used to build JREs that are currently
available.

[#51298625]
[4d93fd2] Exclude debug symbols

Previously, 1.7.0_02 - _05 used a different name for their debug
information and so it was not being filtered from the distributions
properly.  This resulted in these versions of the JRE still being over
100MB in size.  This change ensures that artifacts with these names
are excluded as well.

[#51298625]
[0bf5578] Populated cacerts

Previously, builds of the JREs contained an empty cacerts file.  This
was problematic as it meant that nearly no application that needed to
use SSL would work without some extensive changes.  This change
populates the cacerts file (in all versions of the JRE) with the
certificates found at http://curl.haxx.se/ca/.

[#51298625]
[5765767] Remove .diz files

Previously, the distribution was created including a collection of
.diz files.  These files contain debug information and are unneeded in
our usage.  This change excludes them when create the distribution.

[#51401183]
[62c2bba] Support for OpenJDK 8

This change adds support for OpenJDK 8.  To do this, I toyed with the
idea of running 6, 7, and 8 branches but instead decided to put all of
the logic for the various versions in a single branch.  A decision is
made as to what to execute on the EC2 instance based on the version
being built.
[446ade3] Java 7 support

Added functionality to support building OpenJDK 7.
[a19c31d] Upload to S3

This change uploads the binaries to the proper location in S3 and
updates the index for the uploaded artifacts.
[936128b] Additional dependencies, logging, and upload

This changes adds some additional required dependencies, the
collection of all of the import output to /tmp/build-output.txt, and a
first cut at uploading the binaries to S3.
[e4e593c] Change clone location

Content is now cloned to /tmp rather than $HOME as the script is run
without a reasonable $HOME being specified.
[22e8ec1] Initial commit

This is the initial commit of a tool to automate the creation and
upload of OpenJDK packages.  It consists of a script, run locally,
which starts an EC2 instance.  This instance has a cloud-init
configuration that:

1. installs the required dependencies
2. clones the OpenJDK repository
3. build the binaries

In time it will also upload the binaries to an S3 bucket for download.

Change key structure

As part of the renaming of the JRE bucket, the key structure that the
JREs are upload to, must change.  This change does that work.

[#51641279]

Index contains full URIs

Previously, the index only contained the relative URI to an artifact.
This change ensures that the index now contains the full URI to the
artifact.

[#51641191]

Added note about AWS credentials
Previously, there was no noticed about how to obtain AWS credentials.
This change adds that information.

[#51905291]

Add new JRE

[#52036487]

Change version flag

Previously, the version flag for the builder was -t, which was
confusing.  This change makes it -v and --version.

Update bundle dependencies